refactor: 로또 볼 검증 로직 도메인으로 이동, LottoBallValidator 삭제
