+++ src/main/java/lotto/domain/ticket/ball/LottoBallFactory.java
@@ -7,8 +7,8 @@ import java.util.function.Function;
 import java.util.stream.Collectors;
 import java.util.stream.IntStream;
 
-import static lotto.domain.ticket.ball.LottoBallValidator.LOWER_BOUND;
-import static lotto.domain.ticket.ball.LottoBallValidator.UPPER_BOUND;
+import static lotto.domain.ticket.ball.LottoBall.LOWER_BOUND;
+import static lotto.domain.ticket.ball.LottoBall.UPPER_BOUND;
 
 public class LottoBallFactory {
     private static final Map<Integer, LottoBall> instance;
@@ -24,7 +24,7 @@ public class LottoBallFactory {
     }
 
     public static LottoBall getLottoBallByNumber(int number) {
-        LottoBallValidator.validateNumber(number);
+        LottoBall.validateNumber(number);
         return instance.get(number);
     }
 }
